Consolidate `TestDataProvider`
- Add `TestDataProvider.read`.
- Remove provider from `test_kit`.
- Update tests.
