+++ tests/integration_tests/infrastructure/test_cache_database.py
@@ -19,6 +19,7 @@ from decimal import Decimal
 import pytest
 import redis
 
+from nautilus_trader.backtest.data.providers import TestDataProvider
 from nautilus_trader.backtest.data.providers import TestInstrumentProvider
 from nautilus_trader.backtest.data.wranglers import QuoteTickDataWrangler
 from nautilus_trader.backtest.engine import BacktestEngine
@@ -50,7 +51,6 @@ from nautilus_trader.risk.engine import RiskEngine
 from nautilus_trader.serialization.msgpack.serializer import MsgPackSerializer
 from nautilus_trader.trading.strategy import TradingStrategy
 from tests.test_kit.mocks import MockStrategy
-from tests.test_kit.providers import TestDataProvider
 from tests.test_kit.stubs import TestStubs
 
 
@@ -768,9 +768,10 @@ class TestExecutionCacheWithRedisDatabaseTests:
         self.usdjpy = TestInstrumentProvider.default_fx_ccy("USD/JPY")
 
         wrangler = QuoteTickDataWrangler(self.usdjpy)
+        provider = TestDataProvider()
         ticks = wrangler.process_bar_data(
-            bid_data=TestDataProvider.usdjpy_1min_bid(),
-            ask_data=TestDataProvider.usdjpy_1min_ask(),
+            bid_data=provider.read_csv_bars("fxcm-usdjpy-m1-bid-2013.csv"),
+            ask_data=provider.read_csv_bars("fxcm-usdjpy-m1-ask-2013.csv"),
         )
         self.engine.add_instrument(self.usdjpy)
         self.engine.add_ticks(ticks)
