+++ tests/integration_tests/orderbook/test_orderbook.py
@@ -18,7 +18,6 @@ from nautilus_trader.model.orderbook.book import L1OrderBook
 from nautilus_trader.model.orderbook.book import L2OrderBook
 from nautilus_trader.model.orderbook.book import L3OrderBook
 from nautilus_trader.model.orderbook.error import BookIntegrityError
-from tests.test_kit.providers import TestDataProvider
 from tests.test_kit.stubs import TestStubs
 
 
@@ -32,7 +31,7 @@ def test_l3_feed():
     # immediately, but we may get also delete later.
     skip_deletes = []
     i = 0
-    for i, m in enumerate(TestDataProvider.l3_feed()):  # noqa (B007)
+    for i, m in enumerate(TestStubs.l3_feed()):  # noqa (B007)
         if m["op"] == "update":
             book.update(order=m["order"])
             try:
@@ -64,7 +63,7 @@ def test_l2_feed():
         (68431, "8913f4bf-cc49-4e23-b05d-5eeed948a454"),
     ]
     i = 0
-    for i, m in enumerate(TestDataProvider.l2_feed()):
+    for i, m in enumerate(TestStubs.l2_feed()):
         if not m or m["op"] == "trade":
             pass
         elif (i, m["order"].id) in skip:
@@ -83,7 +82,7 @@ def test_l1_orderbook():
         price_precision=5,
         size_precision=0,
     )
-    for i, m in enumerate(TestDataProvider.l1_feed()):  # noqa (B007)
+    for i, m in enumerate(TestStubs.l1_feed()):  # noqa (B007)
         # print(f"[{i}]", "\n", m, "\n", repr(ob), "\n")
         # print("")
         if m["op"] == "update":
