+++ tests/performance_tests/test_perf_backtest.py
@@ -20,6 +20,7 @@ from decimal import Decimal
 import pandas as pd
 import pytz
 
+from nautilus_trader.backtest.data.providers import TestDataProvider
 from nautilus_trader.backtest.data.providers import TestInstrumentProvider
 from nautilus_trader.backtest.data.wranglers import QuoteTickDataWrangler
 from nautilus_trader.backtest.engine import BacktestEngine
@@ -37,7 +38,6 @@ from nautilus_trader.model.objects import Money
 from nautilus_trader.trading.strategy import TradingStrategy
 from tests.test_kit import PACKAGE_ROOT
 from tests.test_kit.performance import PerformanceHarness
-from tests.test_kit.providers import TestDataProvider
 from tests.test_kit.stubs import TestStubs
 
 
@@ -54,9 +54,10 @@ class TestBacktestEnginePerformance(PerformanceHarness):
 
             # Setup data
             wrangler = QuoteTickDataWrangler(USDJPY_SIM)
+            provider = TestDataProvider()
             ticks = wrangler.process_bar_data(
-                bid_data=TestDataProvider.usdjpy_1min_bid(),
-                ask_data=TestDataProvider.usdjpy_1min_ask(),
+                bid_data=provider.read_csv_bars("fxcm-usdjpy-m1-bid-2013.csv"),
+                ask_data=provider.read_csv_bars("fxcm-usdjpy-m1-ask-2013.csv"),
             )
             engine.add_instrument(USDJPY_SIM)
             engine.add_ticks(ticks)
@@ -89,9 +90,10 @@ class TestBacktestEnginePerformance(PerformanceHarness):
 
             # Setup data
             wrangler = QuoteTickDataWrangler(USDJPY_SIM)
+            provider = TestDataProvider()
             ticks = wrangler.process_bar_data(
-                bid_data=TestDataProvider.usdjpy_1min_bid(),
-                ask_data=TestDataProvider.usdjpy_1min_ask(),
+                bid_data=provider.read_csv_bars("fxcm-usdjpy-m1-bid-2013.csv"),
+                ask_data=provider.read_csv_bars("fxcm-usdjpy-m1-ask-2013.csv"),
             )
             engine.add_instrument(USDJPY_SIM)
             engine.add_ticks(ticks)
@@ -133,9 +135,10 @@ class TestBacktestEnginePerformance(PerformanceHarness):
 
             # Setup data
             wrangler = QuoteTickDataWrangler(USDJPY_SIM)
+            provider = TestDataProvider()
             ticks = wrangler.process_bar_data(
-                bid_data=TestDataProvider.usdjpy_1min_bid(),
-                ask_data=TestDataProvider.usdjpy_1min_ask(),
+                bid_data=provider.read_csv_bars("fxcm-usdjpy-m1-bid-2013.csv"),
+                ask_data=provider.read_csv_bars("fxcm-usdjpy-m1-ask-2013.csv"),
             )
             engine.add_instrument(USDJPY_SIM)
             engine.add_ticks(ticks)
