+++ tests/test_kit/stubs.py
@@ -18,10 +18,12 @@ from datetime import datetime
 from decimal import Decimal
 from typing import List
 
+import orjson
 import pandas as pd
 import pytz
 
 from nautilus_trader.accounting.factory import AccountFactory
+from nautilus_trader.backtest.data.providers import TestDataProvider
 from nautilus_trader.backtest.data.providers import TestInstrumentProvider
 from nautilus_trader.cache.cache import Cache
 from nautilus_trader.common.clock import LiveClock
@@ -32,6 +34,7 @@ from nautilus_trader.common.logging import LiveLogger
 from nautilus_trader.common.logging import LogLevelParser
 from nautilus_trader.core.data import Data
 from nautilus_trader.core.datetime import maybe_dt_to_unix_nanos
+from nautilus_trader.core.datetime import millis_to_nanos
 from nautilus_trader.core.uuid import UUID4
 from nautilus_trader.model.currencies import GBP
 from nautilus_trader.model.currencies import USD
@@ -94,6 +97,7 @@ from nautilus_trader.portfolio.portfolio import Portfolio
 from nautilus_trader.serialization.arrow.serializer import register_parquet
 from nautilus_trader.trading.filters import NewsImpact
 from nautilus_trader.trading.strategy import TradingStrategy
+from tests.test_kit import PACKAGE_ROOT
 from tests.test_kit.mocks import MockLiveDataEngine
 from tests.test_kit.mocks import MockLiveExecutionEngine
 from tests.test_kit.mocks import MockLiveRiskEngine
@@ -874,3 +878,113 @@ class TestStubs:
                 ts_event=maybe_dt_to_unix_nanos(pd.Timestamp(row["Start"])),
                 ts_init=maybe_dt_to_unix_nanos(pd.Timestamp(row["Start"])),
             )
+
+    @staticmethod
+    def l1_feed():
+        provider = TestDataProvider()
+        updates = []
+        for _, row in provider.read_csv_ticks("truefx-usdjpy-ticks.csv").iterrows():
+            for side, order_side in zip(("bid", "ask"), (OrderSide.BUY, OrderSide.SELL)):
+                updates.append(
+                    {
+                        "op": "update",
+                        "order": Order(
+                            price=Price(row[side], precision=6),
+                            size=Quantity(1e9, precision=2),
+                            side=order_side,
+                        ),
+                    }
+                )
+        return updates
+
+    @staticmethod
+    def l2_feed() -> List:
+        def parse_line(d):
+            if "status" in d:
+                return {}
+            elif "close_price" in d:
+                # return {'timestamp': d['remote_timestamp'], "close_price": d['close_price']}
+                return {}
+            if "trade" in d:
+                ts = millis_to_nanos(pd.Timestamp(d["remote_timestamp"]).timestamp())
+                return {
+                    "timestamp": d["remote_timestamp"],
+                    "op": "trade",
+                    "trade": TradeTick(
+                        instrument_id=InstrumentId(Symbol("TEST"), Venue("BETFAIR")),
+                        price=Price(d["trade"]["price"], 4),
+                        size=Quantity(d["trade"]["volume"], 4),
+                        aggressor_side=d["trade"]["side"],
+                        match_id=(d["trade"]["trade_id"]),
+                        ts_event=ts,
+                        ts_init=ts,
+                    ),
+                }
+            elif "level" in d and d["level"]["orders"][0]["volume"] == 0:
+                op = "delete"
+            else:
+                op = "update"
+            order_like = d["level"]["orders"][0] if op != "trade" else d["trade"]
+            return {
+                "timestamp": d["remote_timestamp"],
+                "op": op,
+                "order": Order(
+                    price=Price(order_like["price"], precision=6),
+                    size=Quantity(abs(order_like["volume"]), precision=4),
+                    # Betting sides are reversed
+                    side={2: OrderSide.BUY, 1: OrderSide.SELL}[order_like["side"]],
+                    id=str(order_like["order_id"]),
+                ),
+            }
+
+        return [
+            parse_line(line)
+            for line in orjson.loads(open(PACKAGE_ROOT + "/data/L2_feed.json").read())
+        ]
+
+    @staticmethod
+    def l3_feed():
+        def parser(data):
+            parsed = data
+            if not isinstance(parsed, list):
+                # print(parsed)
+                return
+            elif isinstance(parsed, list):
+                channel, updates = parsed
+                if not isinstance(updates[0], list):
+                    updates = [updates]
+            else:
+                raise KeyError()
+            if isinstance(updates, int):
+                print("Err", updates)
+                return
+            for values in updates:
+                keys = ("order_id", "price", "size")
+                data = dict(zip(keys, values))
+                side = OrderSide.BUY if data["size"] >= 0 else OrderSide.SELL
+                if data["price"] == 0:
+                    yield dict(
+                        op="delete",
+                        order=Order(
+                            price=Price(data["price"], precision=10),
+                            size=Quantity(abs(data["size"]), precision=10),
+                            side=side,
+                            id=str(data["order_id"]),
+                        ),
+                    )
+                else:
+                    yield dict(
+                        op="update",
+                        order=Order(
+                            price=Price(data["price"], precision=10),
+                            size=Quantity(abs(data["size"]), precision=10),
+                            side=side,
+                            id=str(data["order_id"]),
+                        ),
+                    )
+
+        return [
+            msg
+            for data in orjson.loads(open(PACKAGE_ROOT + "/data/L3_feed.json").read())
+            for msg in parser(data)
+        ]
