+++ tests/unit_tests/model/test_model_instrument.py
@@ -17,6 +17,7 @@ from decimal import Decimal
 
 import pytest
 
+from nautilus_trader.backtest.data.providers import TestDataProvider
 from nautilus_trader.backtest.data.providers import TestInstrumentProvider
 from nautilus_trader.model.currencies import AUD
 from nautilus_trader.model.currencies import BTC
@@ -29,14 +30,14 @@ from nautilus_trader.model.objects import Money
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
 from tests.integration_tests.adapters.betfair.test_kit import BetfairTestStubs
-from tests.test_kit.providers import TestDataProvider
 
+
+provider = TestDataProvider()
 
 AUDUSD_SIM = TestInstrumentProvider.default_fx_ccy("AUD/USD")
 USDJPY_SIM = TestInstrumentProvider.default_fx_ccy("USD/JPY")
 XBTUSD_BITMEX = TestInstrumentProvider.xbtusd_bitmex()
 BTCUSDT_BINANCE = TestInstrumentProvider.btcusdt_binance()
-BTCUSDT_BINANCE_INSTRUMENT = TestDataProvider.binance_btcusdt_instrument()
 ETHUSD_BITMEX = TestInstrumentProvider.ethusd_bitmex()
 AAPL_EQUITY = TestInstrumentProvider.aapl_equity()
 ES_FUTURE = TestInstrumentProvider.es_future()
@@ -63,8 +64,9 @@ class TestInstrument:
 
     def test_str_repr_returns_expected(self):
         # Arrange, Act, Assert
-        assert str(BTCUSDT_BINANCE) == BTCUSDT_BINANCE_INSTRUMENT
-        assert repr(BTCUSDT_BINANCE) == BTCUSDT_BINANCE_INSTRUMENT
+        expected = provider.read("binance-btcusdt-instrument-repr.txt").decode()
+        assert str(BTCUSDT_BINANCE) == expected
+        assert repr(BTCUSDT_BINANCE) == expected
 
     def test_hash(self):
         # Arrange, Act, Assert
